requirements.txt
streamlit>=1.29.0
phidata>=2.3.0
groq>=0.4.0
python-dotenv>=1.0.0